Moved the ErrorBoundary component into the components folder.
